packages/gtmcore/requirements.txt
pyyaml==3.12
GitPython==2.1.11
docker==3.4.1
redis==2.10.6
python-redis-lock==3.2.0
rq==0.12.0
jupyter_client==5.2.3
ipykernel==4.8.2
hiredis==0.2.0
requests==2.20.0
rq-scheduler==0.8.3
cryptography==2.3.1
python-jose[cryptography]==3.0.0
schema==0.6.6
blosc==1.5.1
pexpect==4.3.1
natsort==5.1.1
pillow==5.1.0
# click conflicts  7.0 installed by flask-api, 6.7 OK for both
click==6.7
mitmproxy==4.0.4
pandas==0.23.4
aiohttp==3.5.4
aiofiles==0.4.0
python-snappy==0.5.3
packaging==19.0
humanfriendly==4.18
